Refactor credits page
