--- src/components/credits/pages/CreditsPage.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import CreditsContent from '../modules/CreditsContent';
-
-const CreditsPage = () => {
-  return (
-    <>
-      <CreditsContent />
-    </>
-  );
-};
-
-export default CreditsPage;
