+++ src/pages/credits/components/ImageCredits.js
@@ -1,7 +1,9 @@
 import React from 'react';
 import styled from 'styled-components';
+
+// components
 import { Heading, Stack, TextLink } from '@kiwicom/orbit-components/lib';
-import BlackText from '../../text/BlackText';
+import BlackText from '@components/text/BlackText';
 
 const ImageCreditsContainer = styled.div`
   margin-bottom: 50px;
