+++ src/pages/credits/components/index.js
@@ -1,0 +1,5 @@
+import IconCredits from './IconCredits';
+import ImageCredits from './ImageCredits';
+import LibraryCredits from './LibraryCredits';
+
+export { IconCredits, ImageCredits, LibraryCredits };
